Remove masked versions of OXSs, associated methods
OXS (stats) fields don't have a useful notion of wildcards, the openflow
standard reserves the corresponding field.

This commit removes the declarations of the masked OXS fields and
the associated java magic (also making the OXS stack simpler).
