+++ java_gen/java_type.py
@@ -690,15 +690,10 @@ exceptions = {
         'of_oxm_packet_type_masked' : { 'value' : packet_type, 'value_mask' : packet_type },
 
         'of_oxs_byte_count' : { 'value' : u64 },
-        'of_oxs_byte_count_masked' : {'value' : u64, 'value_mask' : u64 },
         'of_oxs_duration' : { 'value' : u64 },
-        'of_oxs_duration_masked' : {'value' : u64, 'value_mask' : u64 },
         'of_oxs_flow_count' : { 'value' : u32obj },
-        'of_oxs_flow_count_masked' : {'value' : u32obj, 'value_mask' : u32obj },
         'of_oxs_idle_time' : { 'value' : u64 },
-        'of_oxs_idle_time_masked' : {'value' : u64, 'value_mask' : u64 },
         'of_oxs_packet_count' : { 'value' : u64 },
-        'of_oxs_packet_count_masked' : {'value' : u64, 'value_mask' : u64 },
 
         'of_oxm_bsn_in_ports_128' : { 'value': port_bitmap_128 },
         'of_oxm_bsn_in_ports_128_masked' : { 'value': port_bitmap_128, 'value_mask': port_bitmap_128 },
